criando um readme simples

+++ src/main/java/com/createpdf/luizdev/model/User.java
@@ -20,5 +20,5 @@ public class User {
     private String name;
     private String email;
     private Date birth;
-    private float height;
+    private Double height;
 }
